==> src/notebooks/1- Explore_and_clean_metadata.ipynb <==
import h5py
import pandas as pd
import numpy as np

metadata = pd.read_csv('../data/raw/MillionSongSubset/SongCSV.csv', index_col='SongNumber')

# Initial data exploration


#Initial data exploration
metadata.info()

ArtistLatitude, ArtistLongitude and SongHotttnesss have a lot of null values
ArtistFamiliarity have a very few.

So let's just drop the columns with a lot of nulls and delete rows with null ArtistFamiliarity values

metadata.drop(labels=['ArtistLatitude', 'ArtistLongitude', 'SongHotttnesss'], axis=1, inplace=True)
metadata = metadata[metadata.ArtistFamiliarity > 0]

metadata

Energy and Danceability look like they are all zeros. Year seems to have a few.

Let's fix this

#Exploración de columnas vacias
print('Empty Danceability values:', metadata[metadata['Danceability'] != 0].shape) #Nothing here, let's delete the column
print('Empty Energy values:', metadata[metadata['Energy'] != 0].shape) #Nothing here, let's delete the column
print('Empty Year values:', metadata[metadata['Year'] != 0].shape) # To many 0 values, let's delete the column

There is no data in Danceability and Energy, and not enough values in Year.

Let's drop all of them.

metadata.drop(labels=['Danceability', 'Energy', 'Year'], axis=1, inplace=True)
metadata

Let's have a look to the rest of the columns, trying to find zeros or empty strings:

print('Empty SongID values:', metadata[metadata['SongID'] == "b''"].shape[0])
print('Empty AlbumID values:', metadata[metadata['AlbumID'] == 0].shape[0])
print('Empty AlbumName values:', metadata[metadata['AlbumName'] == "b''"].shape[0])
print('Empty ArtistID values:', metadata[metadata['ArtistID'] == "b''"].shape[0])
print('Empty ArtistLocation values:', metadata[metadata['ArtistLocation'] == "b''"].shape[0])
print('Empty ArtistName values:', metadata[metadata['ArtistName'] == "b''"].shape[0])
print('Empty Duration values:', metadata[metadata['Duration'] == 0].shape[0])
print('Empty KeySignature values:', metadata[metadata['KeySignature'] == 0].shape[0])
print('Empty KeySignatureConfidence values:', metadata[metadata['KeySignatureConfidence'] == 0].shape[0])
print('Empty Tempo values:', metadata[metadata['Tempo'] == 0].shape[0])
print('Empty TimeSignature values:', metadata[metadata['TimeSignature'] == 0].shape[0])
print('Empty TimeSignatureConfidence values:', metadata[metadata['TimeSignatureConfidence'] == 0].shape[0])
print('Empty Title values:', metadata[metadata['Title'] == "b''"].shape[0])
print('Empty ArtistFamiliarity values:', metadata[metadata['ArtistFamiliarity'] == 0].shape[0])
print('Empty ArtistHotttnesss values:', metadata[metadata['ArtistHotttnesss'] == 0].shape[0])
print('Empty EndOfFadeIn values:', metadata[metadata['EndOfFadeIn'] == 0].shape[0])
print('Empty Loudness values:', metadata[metadata['Loudness'] == 0].shape[0])
print('Empty StartOfFadeOut values:', metadata[metadata['StartOfFadeOut'] == 0].shape[0])

# More cleaning

Confidence columns have a lot of empty values. Let's drop them.
ArtistLocation, KeySignature and EndOfFadeIn also have plenty of empty values.
On top of that, we are going to delete empty records from Tempo, TimeSignature, Title and ArtistHotttnesss.

metadata.drop(labels=
              ['ArtistLocation', 'KeySignature', 'KeySignatureConfidence', 'TimeSignatureConfidence', 'EndOfFadeIn'],
              axis=1, inplace=True)

metadata = metadata[metadata.Tempo > 0]
metadata = metadata[metadata.TimeSignature > 0]
metadata = metadata[metadata.Title != "b''"]
metadata = metadata[metadata.ArtistHotttnesss > 0]

metadata

At a quick glance, we don't see anything wrong with the columns now.

String columns have a strange b at the beggining and simple quotes. Let's remove them.

metadata['SongID'] = metadata['SongID'].apply(lambda x: x[2:-1])
metadata['AlbumName'] = metadata['AlbumName'].apply(lambda x: x[2:-1])
metadata['ArtistID'] = metadata['ArtistID'].apply(lambda x: x[2:-1])
metadata['ArtistName'] = metadata['ArtistName'].apply(lambda x: x[2:-1])
metadata['Title'] = metadata['Title'].apply(lambda x: x[2:-1])

metadata

# Guardamos el dataframe limpio

metadata.to_csv('../data/processed/metadata.csv')
