==> src/notebooks/2- Explore_triplets.ipynb <==
import pandas as pd

#Load from server. It may take long for slow connections.
#location = 'https://projects.danielvivas.com/song_recommender/train_triplets.txt'

#Load from local
location = '../data/raw/TasteProfile/train_triplets.txt'

triplets = pd.read_csv(location, sep='\t', names=['userID', 'songID', 'playCount'])

# Exploring data

This file is pretty simple. Everything is OK here.

triplets

print('Unique users:', len(triplets['userID'].unique()))
print('Unique songs:', len(triplets['songID'].unique()))
print('Total records:', triplets.shape[0])

Just because the file is too big, let's save a sample for slow connections

triplets[:10000].to_csv('../data/raw/TasteProfile/train_triplets_sample.txt')
