==> src/notebooks/3- Prepare_and_analyze_data.ipynb <==
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
%matplotlib inline

# Loading data

metadata = pd.read_csv('../data/processed/metadata.csv', index_col='SongID')


#Load from server. It may take long for slow connections.
#location = 'https://projects.danielvivas.com/song_recommender/train_triplets.txt'

#Sample file for slow connections.
#location = '../data/raw/TasteProfile/train_triplets_sample.txt'

#Load from local
location = '../data/raw/TasteProfile/train_triplets.txt'

triplets = pd.read_csv(location,
                       sep='\t', names=['UserID', 'SongID', 'playCount'], index_col='SongID')

# Operations

Let's merge both datasets. Then, we create a new row with a title - artist string. We'll need it for the recommendation phase.

result = pd.merge(left=metadata, right=triplets, on='SongID')

result['song'] = result['Title'].map(str) + " - " + result['ArtistName']

result.drop(columns= ['SongNumber'])

result

result.to_csv('../data/processed/fulldata.csv')

print('Unique songs in metadata:', metadata.shape[0], '\n')
print('Unique songs in triplets:', len(triplets.index.unique()))
print('Unique users in triplets:', len(triplets['UserID'].unique()), '\n')
print('Unique songs in RESULT:', len(result.index.unique()))
print('Unique users in RESULT:', len(result['UserID'].unique()), '\n')

prueba = triplets[triplets.index.isin(metadata.index.unique())]

print('Unique songs in prueba:', len(prueba.index.unique()))
print('Unique users in prueba:', len(prueba['UserID'].unique()))

len(prueba.index.unique())

Just because the file is too big, let's save a sample for slow connections

result[:10000].to_csv('../data/processed/fulldata_sample.csv')

song_grouped = result.groupby(['song']).agg({'playCount': 'count'}).reset_index()
grouped_sum = song_grouped['playCount'].sum()
song_grouped['percentage']  = song_grouped['playCount'].div(grouped_sum)*100
song_grouped.sort_values(['playCount', 'song'], ascending = [0,1])

analysis = result[['Duration', 'Tempo', 'TimeSignature', 'ArtistFamiliarity', 'Loudness', 'StartOfFadeOut', 'playCount']]

sns.pairplot(analysis)

fig = plt.figure(figsize=(10, 10))
sns.heatmap(result.corr(), annot=True)
fig.show()

sns.scatterplot(data=result[['ArtistFamiliarity', 'playCount']], x='ArtistFamiliarity', y='playCount')

sns.scatterplot(data=result[['Loudness', 'playCount']], x='ArtistFamiliarity', y='playCount')
